+++ frontend/src/components/Header/UserMenu/index.js
@@ -1,0 +1,1 @@
+export { default as UserMenu } from "./UserMenu";
